Add references to the NNPDF code and results
Add the reference of 4.0 paper containing a description of the
hyperoprimization and fitting methodology as well as other potentially
relevant issues such as validation tests or decorrelation.

Add a reference to the NNPDF code, which has been newly made open source
and contains all the data and code required to reproduce the PDF sets,
as well as ancillary analysis.

Closes #80.